pyproject.toml
[tool.poetry]
name = "Secret-Escrow"
version = "0.0.1"
description = ""
authors = ["BOOSTRY Co., Ltd. <dev@boostry.co.jp>"]
license = ""

[tool.poetry.dependencies]
python = "~3.11"
py_ecc = "~7.0.0"
pycryptodome = "~3.20"
web3 = "~6.15.1"
coincurve = "~19.0.1"
eth-brownie = "~1.20.3"
pydantic = "~2.7.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.2.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"
combine_as_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
